+++ tests/plugins.test.ts
@@ -26,6 +26,7 @@ Deno.test("Plugins list in init", () => {
     "postcss",
     "prism",
     "pug",
+    "relations",
     "relative_urls",
     "resolve_urls",
     "sass",