+++ tests/relations.test.ts
@@ -1,0 +1,19 @@
+import { assertSiteSnapshot, build, getSite } from "./utils.ts";
+import relations from "../plugins/relations.ts";
+
+Deno.test("relations plugin", async (t) => {
+  const site = getSite({
+    src: "relations",
+  });
+
+  site.use(relations({
+    foreignKeys: {
+      post: "post_id",
+      category: "category_id",
+      comment: "comment_id",
+    },
+  }));
+
+  await build(site);
+  await assertSiteSnapshot(t, site);
+});